audit(federation): fourth-pass deep audit — fix ~20 security and correctness issues
Security (critical):
- Add OpenSSL::SSL::VERIFY_PEER to all 4 HTTP clients (OAuth,
  PartnerApiClient health probe + execute, WebhookSender) preventing MITM
- Feature gate checks now require explicit boolean true via == true,
  preventing nil/truthy-but-not-boolean values from enabling capabilities
- SQL-parameterise FAILURE_THRESHOLD in record_failure! update_all
- Expand CSV formula injection regex to cover \r and semicolons

Correctness:
- Remove unreachable return-if-blank guard in API key validation
- Fix notification hours calculation: .to_f instead of .to_i prevents
  truncation of fractional second amounts
- Centralise format_federation_hours helpers with i18n suffix, consistent
  negative/zero handling across admin and hub
- Move sanitize_blocked_partner_ids from validate to before_validation
  (validators should not mutate state)
- Round reconciliation run duration to 2dp to prevent float noise
- Add nil-safe platform_type.capitalize in partner show view

Configuration:
- Add maximum caps to ENV config: webhook_timeout ≤300s,
  max_transfer_amount ≤1M, rate_limit ≤10K
- Replace silent rescue nil in engine.rb Sidekiq setup with logged warns

Accessibility:
- Add aria-hidden="true" to decorative SVG in messages empty state

+++ engines/federation_api/app/services/federation/notification_service.rb
@@ -50,7 +50,7 @@ module Federation
         Rails.logger.warn("[Federation::NotificationService] Invalid or missing amount (#{raw_amount.inspect}) for transfer_received notification — defaulting to 0")
         raw_amount = 0
       end
-      hours = (raw_amount.to_i / 3600.0).round(1)
+      hours = (raw_amount.to_f / 3600.0).round(1)
       sender = data[:remote_user_identifier] || I18n.t("federation_mailer.common.default_name")
 
       Federation::NotificationMailer.transfer_received(
@@ -75,7 +75,7 @@ module Federation
         Rails.logger.warn("[Federation::NotificationService] Invalid or missing amount (#{raw_amount.inspect}) for transfer_sent notification — defaulting to 0")
         raw_amount = 0
       end
-      hours = (raw_amount.to_i / 3600.0).round(1)
+      hours = (raw_amount.to_f / 3600.0).round(1)
       recipient = data[:remote_user_identifier] || I18n.t("federation_mailer.common.default_name")
 
       Federation::NotificationMailer.transfer_sent(
